Moved autopilot server to rpcservers folder
