+++ rpcservers/autopilot/config_default.go
@@ -1,7 +1,7 @@
 //go:build !autopilotrpc
 // +build !autopilotrpc
 
-package autopilotrpc
+package autopilot
 
 // Config is empty for non-autopilotrpc builds.
 type Config struct{}
