+++ rpcservers/autopilot/driver.go
@@ -1,7 +1,7 @@
 //go:build autopilotrpc
 // +build autopilotrpc
 
-package autopilotrpc
+package autopilot
 
 import (
 	"fmt"
